+++ infra/infra.go
@@ -1,6 +1,7 @@
 package infra
 
 import (
+	"github.com/alibaba/sealer/infra/aliyun"
 	"github.com/alibaba/sealer/logger"
 	v1 "github.com/alibaba/sealer/types/api/v1"
 )
@@ -13,14 +14,14 @@ type Interface interface {
 
 func NewDefaultProvider(cluster *v1.Cluster) Interface {
 	switch cluster.Spec.Provider {
-	case AliCloud:
-		config := new(Config)
-		err := GetAKSKFromEnv(config)
+	case aliyun.AliCloud:
+		config := new(aliyun.Config)
+		err := aliyun.LoadConfig(config)
 		if err != nil {
 			logger.Error(err)
 			return nil
 		}
-		aliProvider := new(AliProvider)
+		aliProvider := new(aliyun.AliProvider)
 		aliProvider.Config = *config
 		aliProvider.Cluster = cluster
 		err = aliProvider.NewClient()
