+++ infra/utils/utils.go
@@ -1,0 +1,19 @@
+package utils
+
+import (
+	"fmt"
+	"time"
+)
+
+func Retry(tryTimes int, trySleepTime time.Duration, action func() error) error {
+	var err error
+	for i := 0; i < tryTimes; i++ {
+		err = action()
+		if err == nil {
+			return nil
+		}
+
+		time.Sleep(trySleepTime * time.Duration(2*i+1))
+	}
+	return fmt.Errorf("retry action timeout: %v", err)
+}
